requirements.txt
google-cloud-datastore
ipython
requests
